Add check for gridfile generated by new versions of Hypnotoad
Versions of Hypnotoad newer than 1.1.3 generate output that can take
into account whether "shifted" or "identity" parallel transforms are
used. The option should be set to generate a grid file for "shifted" to
be compatible with Hermes. In this case the curvature vector is
generated by Hypnotoad with no contribution from integrated shear, so
the correction applied by Hermes to remove the integrated shear
contribution from the curvature vector should be skipped. This commit
adds a check of the "parallel_transform" variable in the grid file (if
not present indicates that an old enough version of Hypnotoad was used
that the correction is needed), and applies the correction if the grid
file was generated for 'parallel_transform == "identity"', skips it if
'parallel_transform == "shifted"' and throws an exception due to an
unrecognized value for 'parallel_transform' otherwise.