Change how phi boundary is set
The INVERT_SET flag takes the value from the guard cell,
then sets this value on the boundary between cells. This results
in an unstable solution if the guard cells are then put back
in as the new boundary value.

With this fix, the old value of the boundary is taken from the
midpoint, and then put into the guard cell. Hopefully this is
consistent with how the value is used in the Laplacian inversion.