Fix phi in outer boundary cell
I don't think this should be necessary, but it seems to be.
Without setting phi in the outermost boundary cells (+2, -2)
the time advance eventually fails to converge. When this failure
occurs all fields seem to be well behaved except these outer boundary
cells which contain large (1e200) values.

With this change no such failures seem to happen (so far)